backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
python-socketio==5.11.2
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# ── Database ──────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.15.1
psycopg2-binary==2.9.9

# ── Celery Task Queue ────────────────────────────────────────
celery[redis]==5.4.0
redis==5.0.4

# ── Authentication ────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ── Satellite & Geospatial ────────────────────────────────────
sentinelsat==1.2.1
eodag==2.12.1
rasterio==1.3.10
shapely==2.0.4
geopandas==0.14.4
pyproj==3.6.1
fiona==1.9.6

# ── AI / ML ──────────────────────────────────────────────────
ultralytics==8.2.2
torch==2.3.0
torchvision==0.18.0
timm==1.0.3
opencv-python-headless==4.9.0.80
scikit-image==0.23.2

# ── Gemini AI ─────────────────────────────────────────────────
google-generativeai==0.5.4

# ── Browser Automation ────────────────────────────────────────
playwright==1.44.0

# ── Weather ───────────────────────────────────────────────────
herbie-data==2024.3.0

# ── PDF Parsing ───────────────────────────────────────────────
pdfplumber==0.11.0
camelot-py[cv]==0.11.0

# ── Object Storage ────────────────────────────────────────────
minio==7.2.7

# ── Monitoring & Logging ──────────────────────────────────────
structlog==24.1.0
prometheus-fastapi-instrumentator==7.0.0

# ── MLflow ────────────────────────────────────────────────────
mlflow==2.13.0

# ── Google Earth Engine ───────────────────────────────────────
earthengine-api==0.1.403

# ── Utilities ─────────────────────────────────────────────────
httpx==0.27.0
aiofiles==23.2.1
python-dotenv==1.0.1
numpy==1.26.4
Pillow==10.3.0
